requirements.txt
Flask==3.0.2
python-dotenv==1.0.1
groq==0.5.0
playwright==1.44.0
requests==2.31.0
beautifulsoup4==4.12.3
